requirements.txt
alembic==0.7.5.post2
blinker==1.3
cffi==1.5.0
coverage==3.7.1
feedparser==5.2.1
Flask==0.10.1
Flask-Bcrypt==0.6.2
Flask-Bootstrap==3.3.2.1
Flask-DebugToolbar==0.9.2
Flask-Login==0.2.11
Flask-Migrate==1.3.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-Testing==0.4.2
Flask-WTF==0.11
gunicorn==19.4.5
itsdangerous==0.24
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
pycparser==2.14
PyNaCl==1.0.1
python-bcrypt==0.3.1
requests==2.9.1
six==1.10.0
SQLAlchemy==0.9.9
Werkzeug==0.10.4
wheel==0.26.0
WTForms==2.0.2
